+++ components/popover/popover.test.js
@@ -1,9 +1,7 @@
 
 import { createLocalVue, mount } from '@vue/test-utils';
 import DtPopover from './popover.vue';
-import SrOnlyCloseButton from '../../common/sr_only_close_button';
-import axe from 'axe-core';
-import configA11y from '../../storybook/scripts/storybook-a11y-test.config';
+import SrOnlyCloseButton from '../../common/sr_only_close_button.vue';
 import {
   itBehavesLikeVisuallyHiddenCloseButtonExists,
   itBehavesLikeVisuallyHiddenCloseLabelIsNull,
@@ -85,8 +83,8 @@ describe('DtPopover Tests', () => {
   beforeAll(() => {
     // RequestAnimationFrame and cancelAnimationFrame are undefined in the scope
     // Need to mock them to avoid error
-    global.requestAnimationFrame = jest.fn();
-    global.cancelAnimationFrame = jest.fn();
+    global.requestAnimationFrame = vi.fn();
+    global.cancelAnimationFrame = vi.fn();
     testContext.localVue = createLocalVue();
   });
 
@@ -184,7 +182,7 @@ describe('DtPopover Tests', () => {
     describe('When initialFocusElement is none', () => {
       let consoleErrorSpy;
       beforeEach(async () => {
-        consoleErrorSpy = jest.spyOn(console, 'error').mockClear();
+        consoleErrorSpy = vi.spyOn(console, 'error').mockClear();
         await wrapper.setProps({ initialFocusElement: 'none' });
       });
 
@@ -408,15 +406,6 @@ describe('DtPopover Tests', () => {
           expect(popoverWindow.attributes('aria-labelledby')).toBe(wrapper.vm.labelledBy);
         },
       );
-
-      it('should pass axe-core accessibility rules', async () => {
-        const a11yResults = await axe.run(wrapper.element, configA11y);
-        const violations = a11yResults.violations;
-        if (violations.length) {
-          console.log('axe-core accessibility violations:', violations);
-        }
-        expect(violations.length).toEqual(0);
-      });
     });
 
     describe('When popover is closed', () => {